--- a/depends/conanfile_orca.txt
+++ b/depends/conanfile_orca.txt
@@ -1,5 +1,5 @@
 [requires]
-orca/v2.67.0@gpdb/stable
+orca/v2.68.0@gpdb/stable
 
 [imports]
 include, * -> build/include