--- a/src/include/nodes/plannodes.h
+++ b/src/include/nodes/plannodes.h
@@ -456,7 +456,8 @@ typedef Scan TableScan;
 typedef enum LogicalIndexType
 {
 	INDTYPE_BTREE = 0,
-	INDTYPE_BITMAP = 1
+	INDTYPE_BITMAP = 1,
+	INDTYPE_GIST = 2
 } LogicalIndexType;
 
 typedef struct LogicalIndexInfo