--- a/src/test/regress/expected/qp_gist_indexes3_optimizer.out
+++ b/src/test/regress/expected/qp_gist_indexes3_optimizer.out
@@ -147,15 +147,15 @@ SELECT id, property AS "Property" FROM GistTable3
 
 EXPLAIN SELECT id, property AS "Property" FROM GistTable3
  WHERE property ~= '( (999,999), (998,998) )';
-                                    QUERY PLAN                                    
-----------------------------------------------------------------------------------
- Gather Motion 3:1  (slice1; segments: 3)  (cost=100.29..346.42 rows=5 width=36)
-   ->  Bitmap Heap Scan on gisttable3  (cost=100.29..346.42 rows=2 width=36)
+                                   QUERY PLAN                                   
+--------------------------------------------------------------------------------
+ Gather Motion 3:1  (slice1; segments: 3)  (cost=0.00..0.00 rows=1 width=36)
+   ->  Bitmap Table Scan on gisttable3  (cost=0.00..0.00 rows=1 width=36)
          Recheck Cond: property ~= '(999,999),(998,998)'::box
-         ->  Bitmap Index Scan on gistindex3a  (cost=0.00..100.29 rows=2 width=0)
+         ->  Bitmap Index Scan on gistindex3a  (cost=0.00..0.00 rows=0 width=0)
                Index Cond: property ~= '(999,999),(998,998)'::box
  Settings:  enable_seqscan=off
- Optimizer status: legacy query optimizer
+ Optimizer status: PQO version 2.64.0
 (7 rows)
 
 VACUUM ANALYZE GistTable3;
@@ -168,14 +168,15 @@ SELECT id, property AS "ProperTee" FROM GistTable3
 
 EXPLAIN SELECT id, property AS "ProperTee" FROM GistTable3
  WHERE property ~= '( (999,999), (998,998) )';
-                                      QUERY PLAN                                       
----------------------------------------------------------------------------------------
- Gather Motion 3:1  (slice1; segments: 3)  (cost=0.00..200.28 rows=1 width=36)
-   ->  Index Scan using gistindex3a on gisttable3  (cost=0.00..200.28 rows=1 width=36)
+                                        QUERY PLAN                                        
+------------------------------------------------------------------------------------------
+ Gather Motion 3:1  (slice1; segments: 3)  (cost=0.00..3201.22 rows=1600 width=36)
+   ->  Index Scan using gistindex3a on gisttable3  (cost=0.00..3201.01 rows=534 width=36)
          Index Cond: property ~= '(999,999),(998,998)'::box
+         Filter: property ~= '(999,999),(998,998)'::box
  Settings:  enable_seqscan=off
- Optimizer status: legacy query optimizer
-(5 rows)
+ Optimizer status: PQO version 2.64.0
+(6 rows)
 
 -- ----------------------------------------------------------------------
 -- Test: test07Reindex.sql
@@ -207,14 +208,15 @@ SELECT id, property AS "Property" FROM GistTable3
 
 EXPLAIN SELECT id, property AS "Property" FROM GistTable3
  WHERE property ~= '( (999,999), (998,998) )';
-                                      QUERY PLAN                                       
----------------------------------------------------------------------------------------
- Gather Motion 3:1  (slice1; segments: 3)  (cost=0.00..200.27 rows=1 width=36)
-   ->  Index Scan using gistindex3a on gisttable3  (cost=0.00..200.27 rows=1 width=36)
+                                        QUERY PLAN                                        
+------------------------------------------------------------------------------------------
+ Gather Motion 3:1  (slice1; segments: 3)  (cost=0.00..3201.22 rows=1600 width=36)
+   ->  Index Scan using gistindex3a on gisttable3  (cost=0.00..3201.01 rows=534 width=36)
          Index Cond: property ~= '(999,999),(998,998)'::box
+         Filter: property ~= '(999,999),(998,998)'::box
  Settings:  enable_seqscan=off
- Optimizer status: legacy query optimizer
-(5 rows)
+ Optimizer status: PQO version 2.64.0
+(6 rows)
 
 -- ----------------------------------------------------------------------
 -- Test: test08UniqueAndPKey.sql
@@ -233,6 +235,8 @@ EXPLAIN SELECT id, property AS "Property" FROM GistTable3
 --     to create unique GiST indexes or pimary keys on gemoetric data types.
 ------------------------------------------------------------------------------
 CREATE TABLE gisttable_pktest (id integer, property box, poli polygon, bullseye point);
+NOTICE:  Table doesn't have 'DISTRIBUTED BY' clause -- Using column named 'id' as the Greenplum Database data distribution key for this table.
+HINT:  The 'DISTRIBUTED BY' clause determines the distribution of data. Make sure column(s) chosen are the optimal data distribution key to minimize skew.
 CREATE UNIQUE INDEX ShouldNotExist ON gisttable_pktest USING GiST (property);
 ERROR:  access method "gist" does not support unique indexes
 CREATE UNIQUE INDEX ShouldNotExist ON gisttable_pktest USING GiST (poli);