--- a/src/test/regress/init_file
+++ b/src/test/regress/init_file
@@ -12,6 +12,7 @@ m/^NOTICE:  extension "gp_inject_fault" already exists, skipping/
 m/^WARNING:  gpmon:.*Connection refused.*/
 
 m/^ Optimizer status:.*/
+m/^ Settings:.*/
 
 # There are a number of NOTICE and HINT messages around table distribution,
 # for example to inform the user that the database will pick a particular