… src/test/regress/expected/qp_gist_indexes4.out
… src/test/regress/expected/qp_gist_indexes4_optimizer.out
… src/test/regress/input/qp_gist_indexes2.source
… src/test/regress/output/qp_gist_indexes2.source
… src/test/regress/output/qp_gist_indexes2_optimizer.source